15182 Abort update if incompatible user IDs are found
In Checkmk 2.2.0 the use of certain special characters in user IDs is no
longer allowed. As a result, some user IDs that were previously valid in
Checkmk version 2.1.0 will become invalid in version 2.2.0. The
validation affects both local users and users from external connections
such as LDAP. However, local users created in the Web UI are guaranteed
to be compatible as the same restrictions were already applied there.

Since Checkmk 2.2.0 will not work if any incompatible user ID is
present, omd update will look for them and abort if any incompatible
user ID is found.

To be compatible with Checkmk version 2.2.0 please ensure that all user
IDs adhere to the following restrictions: The first character of a user
ID must be a digit, a letter, an underscore (_), or a dollar sign ($).
The rest of the user ID may additionally contain dash (-), at (@), and
dot (.).

CMK-11920

Change-Id: I71297f1fbfdfef0d4be9363be682196aebda133f