CMK-11295: Small refactoring in CSR signing
Change internal representation of validity period from datetime.timedelta
to dateutil.relativedelta.relativedelta.
This has the advantage that it's possible to specify the validity
in months/years without having to worry about the corresponding
amount of days.

This is a preparation step for making the validity period of
agent controller certificates configurable.

Change-Id: I3f2cae9d0b8a4b793eade2df2aab041fa81f5fcd