Don't strip our Livestatus library.
To get usable stack traces, we intentionally don't strip the NEB and the
CMC, so we shouldn't strip our new Livestatus library either: Lots of code
from the former two binaries will end up there.

Removed a useless special case on the way: We never strip the system's
strip, so there is no need to check for that.

Change-Id: I82ed33e2159b2a6d9b3727f340901f7338007238