Moved interface implementations into headers.
At the current point in time, the Livestatus interface will change quite
often.  So after a change in livestatus/Interface.h, *4* files had to be
changed, which was quite annoying.  Moving the implementations to the
headers cuts down the changes by half, which is quite a relief.

The implementations should be tiny, anyway, otherwise we would be doing
something wrong: They should just be tiny views of the underlying data
structures.

And finally: If the compiler can figure out enough types at compilation
time, it can inline the abstractions, perhaps even into one another.  This
is probably not very relevant, but if we are very lucky, this cuts down the
cost of the abstractions a bit.

Perhaps we can even merge all the interface implementations into a single
header like InterfaceImplNeb.h, but that might be too much.  In the end it's
all a balancing act between readability, dependencies and maintainability.
Let's see...

Change-Id: I13e638d7e86ddb3e25ca8ce22356d584c41f2744