mk_oracle: don't remove output of another instance
if running multiple instances of the check (e.g. from two check_mk
servers) at the same time, it leads to the problem that the output will
be removed from the instance that comes last. This leads to the problem
that the oracle instance check will fail.

Fix is to remove only the files that were created by the process itself.