14689 FIX Fixed various Livestatus columns in the NEB
Accessing some Livestatus columns via the NEB could potentially lead to a crash
of the Nagios core. This affects the "service_period" column in the "hosts" or
"services" tables, or the "filename" column in the "hosts" table. This has been
fixed.

If you are running the CMC as the monitoring core, you have not been affected by
this bug.

Change-Id: I1c813895c1b38c6896e2fcd3adc3335a970f3606