13443 FIX list rules endpoint failed to list rulesets containing a hyphen
In previous versions the ruleset name were sanitized, so no escaping was
necessary when building URLs (?)
This was removed with Change-Id
I1a85858fc8881f416f96f4b3f069f558d896b844
But not all occurrences of the replacement were removed, this commit
removes one of them. Hopefully the last one.

There is `static_checks:fileinfo-groups` which could not be found via
REST-API call, that contains a "legal" hyphen that should not be
replaced.

Change-Id: I6cdd5ef305dcb11299a874443f0863b838e6bf33