14724 Improve EC performance when processing large logwatch.ec spool files
The performance and stability of the event console is now improved when
processing large logwatch.ec generated spool files.

logwatch.ec used to be able to generate spool files with unlimited size.

The size of the spool files is now limited to 1 MiB. In case this limit
is reached the multiple spool files are created.

CMK-11808

Change-Id: I803d711277a8b72ffd56b875b4a8a6a10c6654f7