Fixed offset calculation for various columns, fixing lying types.
Due to our type-unaware ColumnOffsets "technology", we had incorrect types
at several places which could not be detected as such by the compiler.  In
an attempt to fix type errors resulting from those incorrect types, an
incorrect fix/werk was attempted, leading to crashes.

The real fix is to fix the offset calculations.  An even better fix would be
making our ColumnOffsets stuff used for joining tables really typed, without
going via intermediate void*.  But with the current code structure, this is
not trivial.  :-/

Change-Id: I12f4c7f69335690a3f4aeb5b8bb31149c0eecc83