13971 FIX Host Tag can now be "null" in REST API responses
None is a valid Tag choice ID and indicated that when the tag was created
the choice ID was left blank. This is probably to accommodate tags with
only one choice, which are displayed as checkboxes in the UI.

This change was inspired by the change with id Idf793dca01755d5d2d27745465f8df6381693a8e
but as 2.0.0 is old stable, I opted for a smaller change to the code.

SUP-12232

Change-Id: I1416be009bc3a91a3c909e6f145be96d50a290a2