Reduce questions for changed file permissions
With Werk #15062 (`eba6aa9c07eab4deac630655bd5bff38a46ce3a2`) the umask
and skel permissions were changed. So in existing installations a
upgrade results in a lot of questions regarding to keep or change
existing file permissions.

If the skel permissions were changed the user was asked for every
change. Now we only ask if the user has changed the permissions from the
previous default. A log will be printed if permissions are changed, not
sure if all users read them properly. But most of the users probably are
not affected so asking them if it is OK to change the permissions if
they don't even know what that means is worse.

CMK-12090

Change-Id: I3a9d6e7f6ac46bea6f33229af23ea512c07ce1e8