FIX 15209 REST-API host_config: fix host attribute IPv6 address validation
Prior to this werk, the REST API failed with a 500 error if the
user tried to query or create a host with a valid Host- or DNS
name as Ipv6 address.

This restriction also caused an error in the DCD as it is reliant
on the the list all hosts endpoint.

This werk resolves both issues.

Change-Id: I0faf2948f455980335daf98e85cb75554b8cc1e6