15251 FIX Fixed "downtimes_with_extra_info" columns in host/service tables
The "is_pending" part of the "downtimes_with_extra_info" column in the
"hosts" and "services" table reported all downtimes to be pending.  This has
been fixed and the information now reflects the real state of the downtime.

This affected only Livestatus for the Nagios core, the CMC part was correct.

Change-Id: I9f44108ed4d4a18c7b7c4727cec64eb15482c185