14581 AWS: Allow configurable piggyback names
When creating a piggyback host for an EC2 instance, <tt>agent_aws</tt> uses the format
C+:
{Private IPv4 address}-{region}-{Instance ID}
C-:
This format guarantees that the piggback name is always unique, even when instances are restarted
and could potentially switch their IP address. However, in practice this problem can be neglected
and only using the IPv4 address as a piggyback name is safe. With this werk, the option
<tt>Piggyback names</tt> in the rule <tt>Amazon Web Services (AWS)</tt> allows you to configure how
piggyback hosts are named.

Note, that if a host changes it's name, then all historical data is lost.

CMK-11372

Change-Id: Ia33ee07e1bd2fa716452844a9c1348d0252fa1d5