update cost query
In the cost table we also have entries for discounts. The sustained use
discounts for example to not have a project id assigned to them. The
updated query filters them out.

The query also includes the project id now, as the name might not be
unique. Updates to the check plugin will come later.

Change-Id: Ibf41d0da106d7214c3cee07caf431429e550f76b