15213 FIX rest_api: time_period exception/active time_ranges default
When creating a time_period via the REST-API, the exception time_ranges
field is not required, however if it's not provided, it will raise a
key error. This werk adds an empty list as a default value, which
aligns with the interface functionality of allowing an empty time_range
value. The second fix is the same issue but for the active time_ranges
field.

CMK-12234
CMK-12181

Change-Id: I936d4f0f312afcf6af99d5b40da8ef32495b95a2