15187 FIX Enforce password policy in REST API and user management
Prior to this Werk both the REST API and the user management UI (Setup > Users) did not correctly enforce the password policy for local accounts.

As a result, administrators with the "User management" permission could set passwords that don't comply with the policy for their own or other users' accounts.

Note that the "Change password" option in the user profile menu was not affected by the issue and correctly checked the password policy.

Change-Id: I24ef58dddb8142ac83cb113ec69b55570c5d83e7