Fixes yet another RPATH related problem
For some reason Bazel:configure_make writes /usr/lib64 to RPATH on
SLES-systems (which is valid there) and for some reasons for dynamic
libraries `-Wl,--rpath,$ORIGIN/../lib` gets appended rather than prepended
to the linker opts, resulting in /usr/lib64/libcrypto.so being loaded
on SLES-Systems, which is too old on SLES 15sp1 only.
This change will still set `-Wl,--rpath,$ORIGIN/../lib` by Bazel, but
(as a fix) set it again in the Make-process which will hopfully evaporate
with some later fix in `foreign_cc`, making it possible to directly set
RPATH for executables and dynamic libraries.

Change-Id: I81de0c448193a966c15d15d6196cfdda24c0b1ed