15145 FIX agent_azure: Remove caching from the azure_usagedetails section
Usage details section provides yesterday's Azure resource costs used in the "Cost"
check plugin. Until now, the section data was fetched once a day and cached.
The problem occurred because yesterday's costs are still being calculated and
changed in Azure's Cost Management throughout the day.

In order to report the cost data as soon as possible and to be consistent with
the Azure's Cost Management, the caching of the azure_usagedetails section has
been removed.

CMK-11943

Change-Id: I8e3d31dc3714f9bb7fa20b36b8810cb6af46d055