15146 azure_virtualmachines: Remove resource group from summary
Previously, the resource group was shown in a summary of a VM service if VMs
are maped to separate hosts (option 'Map data to the VM itself' in the agent config).

With the werk 15261, the cmk/azure/resource_group:<group> label was added
to all Azure hosts, which made this data redundant.

CMK-12164

Change-Id: I47524b7774fd542ebd365b950a34fe0c3d02b3dd