15382 FIX service: remove non functioning endpoint example from swagger-ui
The "Try out" in Swagger-UI does not work for endpoint examples which make use
of the "query" query parameter. The change must be adapted on the Swagger side.
In the meantime, such endpoints will be removed to avoid any confusion and the user
should refer to the examples in the ReDoc instead.

Change-Id: I5394556db51e10e9533076f28b7e13bd4564d90a