15097 FIX Apache WSGI environment when receiving conflicting information
In some cases the URL calculation used internally in Checkmk could be
confused by a WSGI environment with conflicting data. This is now fixed
in code.

In some cases, in the REST API, wrong URLs could be produced and returned
to the user. When accessing those, a 404 error could have ocurred.

Change-Id: I899b74d9405ed53b1ee49e3ad529c55d87dee690