14725 FIX Crash: Builtin dashlets using custom graph timeranges
When using custom graph timeranges and adjusting the "The last 25 hours" option,
rendering builtin dashlets would result in a crash with the following errormessage:

TypeError('NoneType' object is not subscriptable)

Now the dashlet displays displays an MKUserError in the dashlet itself. In this case
you should adjust the dashlet or global configuration.

CMK-11842

Change-Id: Id80535807de18c24370229116f34ce7a4d330d1f