Fix livestatus crash with AuthUser
CmcCore::find_user would pass a temporary CmcContact as
a constant reference to AuthUser. This left the reference
dangling after the function returned, leading to
segmentation faults.

This patch changes the reference to a unique_ptr, thus clarifying
the ownership and fixing the lifetime of the CmcContact object.

See also: https://herbsutter.com/2020/02/23/references-simply/

CMK-12412

Change-Id: I4704b09e91167ef5701e33bc6c62d5f2adf682e9