Prevent load spikes in CI caused by numpy / protobuf compilation
When numpy or protobuf are compiled, they use the number of CPUs to
limit the number of concurrent compile processes. Since we are already
working in parallel across all distro environments, this results in too
much load on the system.

This seems to be the reason for the nightly composition test load spikes
we saw during startup of the tests, where container images and python
virtual environments are built.

Change-Id: Ia5b7dc31360c85f85b3f18e99b97c68adccd3dd1