14301 FIX Minor mail notification usability improvements
Some minor usability improvements have been applied to the body of the mail
notifications.

The Host row now displays the alias only in case it is set and is different
from the hostname to reduce redundant information.

The "performance data" (aka metric) row is not displayed anymore by default
because it's considered a detail information that can be viewed in the UI if it
is needed. However, if you prefer to keep this in the notification, you can
configure this in the notification plugin settings.

Change-Id: I64640f8a18e4a2eea71671b0c35d6abee639963f