14583 FIX DCD: Allow attributes to self repair
Previously, the dcd would stop updating and deleting the host configurations, if they were in
an inconsistent state. The following error would be shown:
C+:
k: api_attributes[k] for k in future_attributes["locked_attributes"] + internal_keys
KeyError: 'site'
C-:
With this werk, the dcd will not crash. Instead it assigns the attributes, which are configured by
the connection.

SUP-11300

Change-Id: Ib622123b14173629063ee3b05b252cf58220497a