14302 Agent plugins: Drop 2.6 compatibility
For compatibility with old Linux systems we still maintain the Python 2
compatibility mechanic we introduced with werk #12149.

In previous versions we ensured compatibility with Python 2.6 and 2.7.
Providing support for 2.6 needs more and more effort these days while it's
relevance decreases. For this reason we drop the support for 2.6 with this
release.

If you are still using Python 2.6 on a monitored system, you will have to use
the old agent plugins or switch to Python 2.7 when updating your agent plugins
to Checkmk 2.2.

Change-Id: If961fc404317f7f6fb5e808eede148a78f41d41f