15188 FIX setting empty user passwords via the REST API
Prior to this Werk it was possible to set empty passwords for users via
the REST API.

An empty password is not considered a valid password and now results in
a 400 error.

Note that this only affected the REST API, not the UI's User Management
nor it's Change Password functionality.

CMK-12264

Change-Id: I668a14b960f685b3e35baecfe427ac8ebc60475d