15376 FIX dcd: Fixes 2 issues with phases 1.1 & 2.4 calling incorrectly formed urls.
This werk fixes 2 issues with the dcd that were introduced after making
the rest api url matching stricter.

The first issue was with the Phase 2.4 Activate Changes failing, This was
because of incorrectly formed redirect urls. The urls are now fixed.

The second issue was with a distributed monitoring setup and with
Phase 1.1 Executing on site... being called, again, with an incorrect
url. The url has now been fixed here too.

Change-Id: I4c01c94e420ef8c5f6b0bbf439277172749273a6