CMK-12284: Extend autocompleters to search for id
This changed was inspired by difficulties in
the Grafana plugin. We discovered that there is
not way to reliably resolve ids to titles for an
external app. The ACs where therefore extended to
match on ids.

The metrics AC already matched on id before this
change, if the match is exact.

Change-Id: I034c751ec6fd59317832bb4765452f41f74dbe5f