livestatus_status inventory_checkmk_server: Extend compatibility
The livestatus checks need to be more carfeful with their compatibility
since not only two parts (Checkmk server, agent) are involved. In addition
to that there are also all the sites on the agent host involved which report
their state via the agent.

It is likely that we have old and new sites on a Checkmk server, especially
during migrations. For this reason we try to be more compatible.

Change-Id: I3ab8c5b5c2aeea3b62a7c92956320cf91764ac6e