15278 FIX synology_disk: read disk health status
A Failing disks would report a working file system (as the file system
was still readable) but still needs to be replaced. The check was
extended to also report disk health status which was introduced with DSM
7.1.

Also split the summary, remove duplicated temperature text, and remove
text about temperature levels, as no temperature levels can be defined
for this service.

SUP-13080

Change-Id: Ib683f46806dc0190ea102df5772720974c98941a