14586 gcp_status: Monitor GCP Status
With this werk, it is possible to monitor the GCP Health Dashboard,
available at https://status.cloud.google.com/. Since this
site is publicly available, no authentication is required for this
monitoring.

This feature consists of a special agent, agent_gcp_status, and a new
check, gcp_status. The special agent can be configured via the rule
Google Cloud Platform (GCP) Status.

Change-Id: I31aa028cbb508571f53ef9aaa42732aa09547641