14304 FIX Fix global setting migration during site update
The Checkmk config update routine (cmk-update-config) was not dealing with
global setting migration correctly. Instead of migrating global and site
specific settings during the site update, the values were only migrated
ad-hoc during runtime in some places and could lead to exceptions in other
places.

This fixes a regression introduced with #14473 in 2.1.0p11.

Change-Id: Ie204565b97929c22b9f95fdb5dad5185d89decd9