14949 FIX Malformed UTF-8 data no longer result in loss of state file
Previously, if for some reason a plugin sent a malformed UTF-8 string,
it broke the integrity of state.pb and resulted in the loss of the state
data. This behavior is explained by the fact that the Protobuf API
allows to write incorrect UTF-8 strings, but does not allow to read them.

This issue has been resolved in this release.

Change-Id: Ie7d79510c8ff99833d2b21092096ca235b5f0a09