14949 FIX Malformed UTF-8 data no longer result in loss of state file
Previously, if for some reason a plugin sent a malformed UTF-8 string,
it broke the integrity of state.pb and resulted in the loss of the state
data. This behavior is explained by the fact that the Protobuf API
allows to write incorrect UTF-8 strings, but does not allow to read them.

This issue has been resolved in this release.

Change-Id: Iad7856369ee6e8f14c024e4daf59f0e897448a10