14728 Netapp Volume Efficiency
The Netapp Special Agent now monitors the logical used space of a volume.

The logical used space includes unlike the physically used space the space savings
due to deduplication, compression and compaction.

For the saved space and logical used space metrics are generated which are
displayed in a new graph called <b>Space savings</b>.

To be able to use this feature the space enforcement of Netapp cannot be set to logical.

CMK-11642

Change-Id: I063dc83e7381c3906823fb94bda916a9638b4f66