15154 FIX fileinfo: Show size in a human readable format
With a migration to the new check API in version 2.1,
fileinfo and fileinfo_group check plugin started showing
size exclusively in bytes.
Before the file size was shown in scaled units (e.g. KiB, MiB, GiB, ...).

With this werk, file size is shown in a human readable format again.

SUP-13474

Change-Id: I1668c8bec0aff3a0958f9f79b8cc4f870207eea9