fix typing for toggle_all_rows
after adding types to toggle_all_row's signature where the
first parameter has to be always set. I figured out that
this isn't always the case which causes some error and
changes the behavior of our code, for this reason I make
the first parameter optional to match the original behavior.

Change-Id: I4454cf160460bb3cb7b4d956d333b9783aaef536