Simplify data_provider #2
Inline function with a single caller.  We can also safely remove the
corresponding tests as the tests for `get_parsed_section()` were
redundant with the tests for `get_section_kwargs()`.

It is also noteworthy that the signature of `get_parsed_section()`
was different and confusing with regard to the signatures of the
other two functions on ParsedSectionsBroker.  Also note that the
signature of `get_section_kwargs()` _is consistent_ with the
signatures of `get_cache_info()` and `filter_available()`.

Change-Id: Ib1cd7ac8a6b72a4fc6772da95bb2c52aa4f2c9ff