15281 FIX mssql_blocked_sessions: Levels now overwrite state
The inline-help of the parameter "Levels for wait" of the rule "MSSQL
Blocked Sessions" mentions:

"The thresholds for wait_duration_ms. Will overwrite the default state
set above."

This was not right. If only the levels were set, the check would go to
CRIT regardless of the wait_duration_ms.

After setting "Levels for wait" all sessions that wait for a shorter
duration than the warning level of "Levels for wait" will not affect the
check result. The check result is then controlled by wait_duration_ms
being greater than the levels defined.

If "Levels for wait" is not set (the default), the behaviour is not
altered.

SUP-13208

Change-Id: I16e557e2319ba99eefb1a982b243883d8f64e0fd