FIX 15153 azure_virtual_machine: Redesign service rule
Previously, if a mapping for power or provisioning states was configured for some,
but not for all states, the unconfigured states would always be mapped to WARN.
This behavior was confusing and forced the users to configure mapping for all
the states even when they wanted to change the behavior of only one. Moreover,
since the parameters of the rule used to be groups of states, the rule matching
could not take place at the state level.

With this werk, if you configure the state mapping for a single state the state
mapping of all other states will be unaffected.

CMK-12523

Change-Id: I0e7ae45a921193e266e4b248cf29f2d3f0cfb05a