Remove SectionParser tuple
Firing a new abstraction for a two-tuple isn't really worth it.
Moreover, handling the name separately from the business logic is
advantageous in many cases as they tend to follow different code
paths---as can be seen here as well where most of the logic is
performed on the name only.

See also:
   https://wiki.haskell.org/Fairbairn_threshold

Change-Id: I63b9ab4a5a4e84c2f3a41427ec11405d05a3dd96