Added version/compatibility assertions
MIN_VERSION added to the BaseVersion class to specify the minimum
version for a fully supported update which can merge the config.
The environment variable "MIN_VERSION" can be exported in the test
environment to override the default value in conftest.py.

Based on the value of MIN_VERSION, trying to update lower versions
will be expected to fail while updating equal or higher versions
will be expected to pass. This way we can do positive and negative
checks in one go with ease.

Change-Id: I69b09a40d30bcb55bf875e22279e6e2586314b80