Instrument tests with helgrind
We've had instability in the "composition tests" for some time now.
They could be tracked down to a crash where the CMC dumped core.
However, analysis of the core did not help finding the source of the
crash.  The crash could not be reproduced locally either, not under load
or killing/restarting the CMC in tight loops.

We therefore resort here to instrumenting the tests themselves, this
time with helgrind.

CMK-12511

Change-Id: I25824ab547177cb7212a6f3c61ef1282a9000678