14950 FIX Correct processing of the rule Install Python runtime environment
Previously, even if the rule "Installation" had been set to
"Never install Python", the "Python environment" rule might be
used by WATO when creating MSI file even though it contradicts
the "Never install" instruction.
If the rule "Python environment" is set in "Deploy standard
environment", this results in the inability to use the MSI
installer on older operating system as Windows Vista and Windows
Server 2008 R1 without a clear reason.
This affected all systems with which Standard Python environment
(Python 3.8 based) is incompatible.

With this release, the problem has been fixed

Change-Id: Id0c984e3a6ad2d706381a2386e96a6360b0135b2