gcp_status: remove time dependence
The check only shows the incidents, which are currently affected,
and therefore they should always be open. Thus, we can omit the check.
We could have considered adding incidents, which are inactive. However,
The user can use the Alert Statistics to see incidents, which affected
them in the past. Moreover, the URI of the incident provides an
overview of each incident and its resolution.

Moreover, showing locations, if everything is grouped by location
does not make a lot of sense. Thus, we omit this information also.

Change-Id: If6c9621a5eeb2dbd89fb51d57492aeaa27219eb7