REST_API: fix auth_type in user_config
Before this change it was assumed that enforce_pw_change is an indicator
that the user is using auth_type=password, but this seems not to be
right.
I checked the first unittest adaption with a running site. If you use
the payload of the unittest to create a user, the site is showing a user
with Authentication=none which can not be edited via the UI.

Change-Id: Ia2b058c56c29f7330dbe936bf1a67979aa9a81c8