sources: Go CRIT if missing agent source
A truthy `is_tcp_host` means that the user did *not* select "No API
integrations, no Checkmk agent" in the host configuration.  That implies
that the user expects the Checkmk agent, a special agent, or both.  If
the configuration couldn't find any, then it's an error and should
trigger CRIT.

CMK-12650

Change-Id: I21bba33eabdbca5cb5a2674727bb55ce0ccb92bd