Revert "Check for stored walks with simulation mode"
If simulation is set, the data is always obtained from the
FileCache and never from the Fetcher.

However, simulation of SNMP plugs a StoredWalkSNMPBackend into
the Fetcher (although it really *is* a cache).  So, in that very
case, we actually want to skip the FileCache and obtain the
cached data *from the fetcher*!

Not all simulation modes are born equal.

CMK-12721

This reverts commit 5caa902a32dde84499dc4739394875a02451ac39.

Change-Id: Ie5f3ff3479351276f1af61a671733e22c73be588