15553 FIX time_period: making time-related validation more strict
Prior to this werk, the REST API was quite lenient when it came
to specifying a time allowing the format hh:mm for some while asking
the hh:mm:ss format for others. This werk targets the active_time_ranges
field where now a hh:mm:ss format is enforced.

Due to the introduction of the more strict pattern, user scripts must be
adjusted accordingly to respect the new format.

Change-Id: I97fa4c02f6a874a0146f0fa841e7df7988cdef3e