15459 FIX filestats: Fix crash when monitoring a symlink with a non-existing target
Until now, the check plugin would crash if a monitored file didn't have age and
size information. This case is possible for symlinks with a non-existing target,
files that vanished during agent plugin run and in case of errors.

Now, the affected files will be ignored when calculating age and size extremes and
the status of the file will be shown in case of 'Single file' aggregation.

CMK-12813

Change-Id: I2e849213b70bbb69acad8c4dddf3eb3aba6abf94