Fix up missunderstood version requirement
A required version of "2.2.0" would mean the MKP is not
allowed in "2.2.0b1", as "2.2.0" is the specific version
released after the last beta release, and before the first
patch release.

This function can be removed once "2.2.0" is actually
released. For the future, we plan to implement a dialog
to ask the user whether they really mean to exclude beta
releases.

Change-Id: I3dc94bda633f2fc1d05c017bac24ac02386046ba