15524 Avoid confusion with MKPs 'Minimum required Checkmk version'
This only affects users that create MKPs and make use of the field 'Minimum required Checkmk version'.

This field is usually prepopulated with the version number of the site (e.g. "2.1.0p17").
We found users frequently removed the patch release number (turning it into "2.1.0"), thinking that this tranlates as "any 2.1 version".

However, it does not. "2.1.0" is the specific version released after "2.1.0b9" and before "2.1.0p1".
Consequently, an MKP that requires version 2.2.0, for instance, will not be available in verison 2.1.0b1.

This is most likely not what users wanted, so we now prompt a warning if such a version number is specified.
You can still configure it, though.

Change-Id: If3f074702e6c08ffc19e0058f008b04aa9f71ec8