15070 SEC Drop support for weak DH ciphers
With this Werk two TLS ciphers are disabled from the <tt>stunnel</tt> configuration.
<tt>stunnel</tt> is used when the <i>Encrypt communication</i> option in <i>Enable Livestatus access via network (TCP)</i> or <i>Notification Spooler Configuration</i> is used.

To our knowledge no attacks on these ciphers are known, this is a hardening measure.

We rate this with a CVSS of 0 (None) (CVSS:3.1/AV:N/AC:L/PR:N/UI:N/S:U/C:N/I:N/A:N). This CVSS is primarily meant to please automatic scanners.

Change-Id: I0a098eccc7a90cd62dc156819b09e90003cba2db