Reland: Consolidate integration test execution
Execute integration tests from venv (like all other tests). We now test
the site from the outside and use the Site object to access the
site resources.

This commit will break most integration tests which will be changed
to work in the new mode in the follow up commits.

Following this change we can simplify a number of special cases in the
conftest and testlib.

Change-Id: I7beb6c03ff6c73fc0b4264f2147124f608abeea0