15592 FIX agent: update openapi doc examples for streamed data responses
This werk introduces a change to the openapi documentation for
endpoints that support streamed responses. The examples will
now correctly show that stream=True is required for urllib
and requests examples.  Also, streamed responses reply with a
status code of 200 and not a 204 as was previously shown
in the examples.

CMK-11044

Change-Id: Ib713b36ba6be9a5f876c65663f21111e95e19be5