Let "plugins" handle edition
The modules are loaded explicitly via import and over the plugin-
specific `load_plugins()` magic.  However, the `load_plugins()` magic
didn't respect the edition.  This resulted in plugins loaded with the
wrong edition-specific code.

The edition-specific plugins are now loaded under an edition guard. The
edition guard is independent from the the loading mechanism and is
therefore less fragile.

CMK-12635

Change-Id: I8b5aa03a5ff4f1ac2a927388b0de0d484b6b8e4b