15415 FIX Interface checks: Fix error in perfometer rendering
From 2.2 onwards, the interface services do not produce useless and
potentially wrong all-zero metrics anymore (werk #14542). This had an
unwanted side effect: Until the counters are initialized, there was a
fallback to the legacy perfometer because the non-legacy one is missing
metrics. The legacy perfometer assumed that all metrics are present
(pre-2.2 behaviour), which led to a crash.

To solve this issue, we simply remove the legacy perfometer for
interface checks.

CMK-12785

Change-Id: I2f980823c7161ad70f18590abea99f18c729d829